refactor AddViewer form
